#add module8-task1 - add validate-form.js and witch-capacity.js
