+++ js/main.js
@@ -1,5 +1,6 @@
 import {createObject, createObjects} from './modules/create-object.js';
 import {createAdElement, createAdsFragment} from './modules/similar-ads.js';
+import {validateForm} from './modules/validate-form.js';
 
 const mapCanvas = document.querySelector('#map-canvas');
 const object = createObject();
@@ -9,3 +10,4 @@ const adsFragment = createAdsFragment(objects);
 
 mapCanvas.appendChild(ad);
 mapCanvas.appendChild(adsFragment);
+validateForm();
