+++ js/modules/similar-ads.js
@@ -1,78 +1,7 @@
 import {createObject} from './create-object.js';
-
-const isEmpty = (selector, property) => {
-  if(property === undefined) {
-    selector.classList.add('hidden');
-  } else {
-    selector.classList.remove('hidden');
-    selector.textContent = property;
-  }
-};
-
-const dataSwitching = (obj) => {
-  const templateFragment = document.querySelector('#card').content;
-  const template = templateFragment.querySelector('.popup');
-  const ad = template.cloneNode(true);
-  const adType = ad.querySelector('.popup__type');
-
-  switch(obj.offer.type) {
-    case 'flat':
-      adType.textContent = 'Квартира';
-      break;
-    case 'bungalow':
-      adType.textContent = 'Бунгало';
-      break;
-    case 'house':
-      adType.textContent = 'Дом';
-      break;
-    case 'palace':
-      adType.textContent = 'Дворец';
-      break;
-    case 'hotel':
-      adType.textContent = 'Отель';
-      break;
-  }
-};
-
-const dataChecking = (obj) => {
-  const templateFragment = document.querySelector('#card').content;
-  const template = templateFragment.querySelector('.popup');
-  const ad = template.cloneNode(true);
-  const adTextTime = ad.querySelector('.popup__text--time');
-  const adType = ad.querySelector('.popup__type');
-  const adAvatar = ad.querySelector('.popup__avatar');
-  const adTextCapacity = ad.querySelector('.popup__text--capacity');
-
-  if(obj.offer.type === undefined) {
-    adType.classList.add('hidden');
-  } else {
-    adType.classList.remove('hidden');
-  }
-
-  if(obj.offer.rooms === undefined || obj.offer.guests === undefined) {
-    adTextCapacity.classList.add('hidden');
-  } else {
-    adTextCapacity.classList.remove('hidden');
-  }
-
-  if(obj.offer.checkin === undefined || obj.offer.checkout === undefined) {
-    adTextTime.classList.add('hidden');
-  } else {
-    adTextTime.classList.remove('hidden');
-  }
-
-  if(ad.querySelectorAll('.popup__feature').length === 0) {
-    ad.querySelector('.popup__features').classList.add('hidden');
-  } else {
-    ad.querySelector('.popup__features').classList.remove('hidden');
-  }
-
-  if(obj.author.avatar === undefined) {
-    adAvatar.classList.add('hidden');
-  } else {
-    adAvatar.src = obj.author.avatar;
-  }
-};
+import {isEmpty} from '../utils/is-empty.js';
+import {dataSwitching} from '../utils/data-switching.js';
+import {dataChecking} from '../utils/data-checking.js';
 
 const createAdElement = (object) => {
   const templateFragment = document.querySelector('#card').content;
